tidy up:
 - remove unneeded imports
 - move only function from tools to event_io_file; delete tools
 - move TypeInfo down to where it is used first
